Remove parser dependency
